requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-chess==1.999
openai==1.51.0
python-dotenv==1.0.0
gunicorn==22.0.0
eventlet==0.33.3
httpx==0.27.0
